requirements.txt
python-dotenv
openai
flask
langchain-community
langchain-openai
